@@ -5,8 +5,10 @@
 //  Created by Martin Lukacs on 10/04/2023.
 //
 
+import Foundation
+
 // MARK: - Photos
-public struct Photos: Codable {
+public struct Photos: Codable, Sendable {
     public let photos: [Photo]
 }
 
@@ -15,8 +17,17 @@ public struct Photo: Codable, Identifiable, Equatable, Hashable, Sendable {
     public let id, sol: Int
     public let camera: Camera
     public let imgSrc: String
-    let earthDate: String
-    let rover: Rover
+    public let earthDate: String
+    public let rover: Rover
+
+    public init(id: Int, sol: Int, camera: Camera, imgSrc: String, earthDate: String, rover: Rover) {
+        self.id = id
+        self.sol = sol
+        self.camera = camera
+        self.imgSrc = imgSrc
+        self.earthDate = earthDate
+        self.rover = rover
+    }
 
     enum CodingKeys: String, CodingKey {
         case id, sol, camera
@@ -24,26 +35,21 @@ public struct Photo: Codable, Identifiable, Equatable, Hashable, Sendable {
         case earthDate = "earth_date"
         case rover
     }
+
+    public var imageUrl: URL? {
+        let url = imgSrc.contains("https:") ? imgSrc : imgSrc.replacingOccurrences(of: "http:", with: "https:")
+        return URL(string: url)
+    }
 }
 
+extension Photo: Mockable {
+   static public var mocked: Photo {
+        Photo(id: 10,
+              sol: 1200,
+              camera: Camera.mocked,
+              imgSrc: "https://mars.jpl.nasa.gov/msl-raw-images/proj/msl/redops/ods/surface/sol/00000/opgs/edr/fcam/FRA_397502305EDR_D0010000AUT_04096M_.JPG",
+              earthDate: "2004-02-03",
+              rover: Rover.mocked)
+    }
+}
 
-//
-//{
-//"id": 102693,
-//"sol": 1000,
-//"camera": {
-//"id": 20,
-//"name": "FHAZ",
-//"rover_id": 5,
-//"full_name": "Front Hazard Avoidance Camera"
-//},
-//"img_src": "http://mars.jpl.nasa.gov/msl-raw-images/proj/msl/redops/ods/surface/sol/01000/opgs/edr/fcam/FLB_486265257EDR_F0481570FHAZ00323M_.JPG",
-//"earth_date": "2015-05-30",
-//"rover": {
-//"id": 5,
-//"name": "Curiosity",
-//"landing_date": "2012-08-06",
-//"launch_date": "2011-11-26",
-//"status": "active"
-//}
-//},
