@@ -1,6 +1,6 @@
 //
 //  APIConfiguration.swift
-//  
+//
 //
 //  Created by Martin Lukacs on 10/04/2023.
 //
@@ -8,6 +8,6 @@
 import Foundation
 
 public enum APIConfiguration {
-    static let apiKey = "hauCFuire0xmOm1JVntGzUITXVWTLVK9mIztmIaF"//"YOUR_API_KEY_HERE"
-    public static let baseUrl = "https://api.nasa.gov/mars-photos/api/v1/rovers"
+    static let apiKey = "hauCFuire0xmOm1JVntGzUITXVWTLVK9mIztmIaF" // "YOUR_API_KEY_HERE"
+    public static let baseUrl = "https://api.nasa.gov/mars-photos/api/v1/"
 }
