@@ -1,6 +1,6 @@
 //
 //  MarsMissionEndpoints.swift
-//  
+//
 //
 //  Created by Martin Lukacs on 12/04/2023.
 //
@@ -23,23 +23,25 @@ struct RequestParams: Codable, Hashable {
 
 enum MarsMissionEndpoint {
     case rover(id: String)
+    case manifest(id: String)
     case photoWithSol(request: RequestParams)
     case photoFromDate(request: RequestParams)
 }
 
-extension MarsMissionEndpoint: Endpoint  {
-
+extension MarsMissionEndpoint: Endpoint {
     var baseUrl: String? {
         APIConfiguration.baseUrl
     }
 
     var path: String {
         switch self {
-        case .rover(let id):
-            return "/\(id)"
-        case .photoWithSol(let params),
-                .photoFromDate(let params):
-            return "/\(params.roverId)/photos"
+        case let .rover(id):
+            return "rovers/\(id)"
+        case let .manifest(id):
+            return "manifests/\(id)"
+        case let .photoFromDate(params),
+             let .photoWithSol(params):
+            return "rovers/\(params.roverId)/photos"
         }
     }
 
@@ -47,22 +49,21 @@ extension MarsMissionEndpoint: Endpoint  {
         .get
     }
 
-    var header: [String : String]? {
+    var header: [String: String]? {
         nil
     }
 
     var body: [String: Any]? {
         var params: [String: Any] = ["api_key": APIConfiguration.apiKey]
         switch self {
-        case .photoWithSol(let param), .photoFromDate(let param):
+        case let .photoFromDate(param), let .photoWithSol(param):
             guard let dict = param.convertToDictionary else {
                 return params
             }
-            params = params.merging(dict) { (current, _) in current }
+            params = params.merging(dict) { current, _ in current }
             return params
         default:
             return params
-
         }
     }
 }
