@@ -1,7 +1,7 @@
 //
-//  
+//
 //  GetMarsPhotosBasedOnDate.swift
-//  
+//
 //
 //  Created by Martin Lukacs on 10/04/2023.
 //
@@ -10,11 +10,12 @@
 import DomainInterfaces
 import NasaModels
 
+// sourcery: AutoMockable
 public protocol GetMarsPhotosBasedOnDateUseCase: Sendable {
-   func execute(for roverId: RoverIdentification,
-                at date: String,
-                for camera: String?,
-                and page: Int?) async throws -> [Photo]
+    func execute(for roverId: RoverIdentification,
+                 at date: String,
+                 for camera: String?,
+                 and page: Int?) async throws -> [Photo]
 }
 
 public extension GetMarsPhotosBasedOnDateUseCase {
@@ -31,7 +32,7 @@ public extension GetMarsPhotosBasedOnDateUseCase {
                         at date: String,
                         for camera: String? = nil,
                         and page: Int? = nil) async throws -> [Photo] {
-     try await execute(for: roverId, at: date, for: camera, and: page)
+        try await execute(for: roverId, at: date, for: camera, and: page)
     }
 }
 
@@ -42,14 +43,13 @@ public final class GetMarsPhotosBasedOnDate: GetMarsPhotosBasedOnDateUseCase {
         self.repository = repository
     }
 
-    
     public func execute(for roverId: RoverIdentification,
-                 at date: String,
-                 for camera: String?,
-                 and page: Int?) async throws -> [Photo] {
+                        at date: String,
+                        for camera: String?,
+                        and page: Int?) async throws -> [Photo] {
         try await repository.getPhotosByDate(for: roverId,
-                                                  at: date,
-                                                  for: camera,
-                                                  and: page)
+                                             at: date,
+                                             for: camera,
+                                             and: page)
     }
 }
