@@ -1,36 +1,40 @@
 //
-//  
+//
 //  GetAllOrderedRovers.swift
-//  
+//
 //
 //  Created by Martin Lukacs on 10/04/2023.
 //
 //
 
-import NasaModels
 import Foundation
+import NasaModels
 
+// sourcery: AutoMockable
 public protocol GetAllOrderedRoversUseCase: Sendable {
-   func execute() async throws -> [Rover]
+    func execute() async throws -> [Rover]
 }
 
 public extension GetAllOrderedRoversUseCase {
     func callAsFunction() async throws -> [Rover] {
-      try await execute()
+        try await execute()
     }
 }
 
 public final class GetAllOrderedRovers: GetAllOrderedRoversUseCase {
-    private let getRoverUseCase: GetRoverInformationsUseCase
+    private let getRoverInformations: any GetRoverInformationsUseCase
 
-    public init(getRoverUseCase: GetRoverInformationsUseCase) {
-        self.getRoverUseCase = getRoverUseCase
+    public init(getRoverInformations: any GetRoverInformationsUseCase) {
+        self.getRoverInformations = getRoverInformations
     }
-    
+
     public func execute() async throws -> [Rover] {
-        let curiosityInfos = try await getRoverUseCase(for: .curiosity)
-        let opportunityInfos = try await getRoverUseCase(for: .opportunity)
-        let spiritInfos = try await getRoverUseCase(for: .spirit)
-        return [curiosityInfos, opportunityInfos, spiritInfos].sorted(using: KeyPathComparator(\.name))
+        async let curiosityInfos = getRoverInformations(for: .curiosity)
+        async let opportunityInfos = getRoverInformations(for: .opportunity)
+        async let spiritInfos = getRoverInformations(for: .spirit)
+        async let perseveranceInfos = getRoverInformations(for: .perseverance)
+
+        return try await [curiosityInfos, opportunityInfos, spiritInfos, perseveranceInfos]
+            .sorted(using: KeyPathComparator(\.name))
     }
 }
