@@ -1,7 +1,7 @@
 //
-//  
+//
 //  GetRoverInformations.swift
-//  
+//
 //
 //  Created by Martin Lukacs on 10/04/2023.
 //
@@ -10,6 +10,7 @@
 import DomainInterfaces
 import NasaModels
 
+// sourcery: AutoMockable
 public protocol GetRoverInformationsUseCase: Sendable {
     func execute(for roverId: RoverIdentification) async throws -> Rover
 }
@@ -26,7 +27,7 @@ public final class GetRoverInformations: GetRoverInformationsUseCase {
     public init(repository: MarsMissionInformationsServicing) {
         self.repository = repository
     }
-    
+
     public func execute(for roverId: RoverIdentification) async throws -> Rover {
         try await repository.getInformation(for: roverId)
     }
