@@ -6,6 +6,7 @@
 //
 
 import Foundation
+
 @preconcurrency import Factory
 
 @MainActor
@@ -18,10 +19,14 @@ extension RouterContainer {
     var tabViewRouter: Factory<TabViewRouter> {
         self { TabViewRouter() }
     }
+
+    var mainRouter: Factory<MainPathRouter> {
+        self { MainPathRouter() }
+    }
 }
 
-extension RouterContainer {
-    func autoRegister() {
+extension RouterContainer: AutoRegistering {
+    nonisolated func autoRegister() {
         manager.defaultScope = .singleton
     }
 }
