@@ -9,6 +9,7 @@ import Foundation
 import SwiftUI
 
 // MARK: - TabView Destinations
+
 enum MainTabDestination {
     case missions
     case favorites
@@ -44,12 +45,12 @@ enum MainTabDestination {
 @MainActor
 final class TabViewRouter {
     @ViewBuilder
-    func goToScene(for destination: MainTabDestination) -> some View {
+    func navigate(to destination: MainTabDestination) -> some View {
         switch destination {
         case .missions:
-            Text("Home")
+            MissionView()
         case .favorites:
-            Text("Favorites")
+            FavoriteView()
         }
     }
 }
