@@ -1,5 +1,5 @@
 //
-//  
+//
 //  MainTabViewView.swift
 //  SwiftOnMars
 //
@@ -8,15 +8,23 @@
 //
 
 import Factory
+import SOMDesignSystem
 import SwiftUI
 
 struct MainTabView: View {
+    @Environment(\.scenePhase) private var scenePhase
     @StateObject private var viewModel = MainTabViewModel()
-    @Injected(\RouterContainer.tabViewRouter) private var router
+    private let tabRouter = resolve(\RouterContainer.tabViewRouter)
     @State private var selectedTabId = 0
-    
+
     var body: some View {
         tabView
+            .onChange(of: scenePhase) { newPhase in
+                guard newPhase == .background else {
+                    return
+                }
+                viewModel.persist()
+            }
     }
 }
 
@@ -26,22 +34,24 @@ struct MainTabViewView_Previews: PreviewProvider {
     }
 }
 
-
-private  extension MainTabView {
+private extension MainTabView {
     var tabView: some View {
         TabView(selection: $selectedTabId) {
             createTabItem(for: .missions)
             createTabItem(for: .favorites)
         }
-        .onChange(of: selectedTabId) { id in
-            print("Selected id: \(id)")
+        .onAppear {
+            let appearance = UITabBarAppearance()
+            UITabBar.appearance().scrollEdgeAppearance = appearance
         }
+        .tint(Asset.Colors.Main.primaryDarkest.color)
     }
 }
 
 private extension MainTabView {
+    @ViewBuilder
     func createTabItem(for destination: MainTabDestination) -> some View {
-        router.goToScene(for: destination)
+        tabRouter.navigate(to: destination)
             .tabItem {
                 Label(destination.name, systemImage: destination.icon)
             }
